+++ test/blackbox-tests/test-cases/jsoo/inline-tests.t/run.t
@@ -5,10 +5,10 @@ Run inline tests using node js
   > EOF
 
   $ dune runtest
-           run alias byte/runtest
+  inline_test_runner_inline_tests_byte alias byte/runtest
   inline tests (Byte)
   inline tests (Byte)
-           run alias native/runtest
+  inline_test_runner_inline_tests_native alias native/runtest
   inline tests (Native)
   inline tests (Native)
           node alias js/runtest